Switch from bootstrap-wysihtml5 to wysihtml5, and use bower for client JS
Bower, Rails and Heroku docs are here: https://coderwall.com/p/6bmygq

We were barely using bootstrap-wysihtml5's added features in the first
place, the Bower-based version didn't support Bootstrap 3, and the
bundled copy of wysihtml5 was out of date.  So let's use wysihtml5
directly.

This means losing the "insert link" feature for now, but I wanted to
clean out that that toolbar anyway.

This also means we now use stacked buildpacks on Heroku.  So far, this
actually appears to work.  I'm impressed.